Move account code into AccountSection component
